接続用のインスタンス変数を初回起動時にしか生成しないように設定
