+++ app.rb
@@ -4,6 +4,11 @@ require 'sinatra'
 require 'sinatra/reloader'
 require 'csv'
 require 'securerandom'
+require_relative 'lib/memo'
+
+configure do
+  set :database, MemoDatabase.new
+end
 
 helpers do
   def h(text)
@@ -16,8 +21,7 @@ get '/' do
 end
 
 get '/memos' do
-  database = MemoDatabase.new
-  @memos = database.load_all_memos
+  @memos = settings.database.load_all_memos
   @title = 'top'
 
   status 200
@@ -32,8 +36,7 @@ get '/memos/new' do
 end
 
 get '/memos/:memo_id' do
-  database = MemoDatabase.new
-  @memo = database.search_memo_by_id(params['memo_id'])
+  @memo = settings.database.search_memo_by_id(params['memo_id'])
   pass if @memo.nil?
   @id = params['memo_id']
   @title = 'Show memo'
@@ -43,8 +46,7 @@ get '/memos/:memo_id' do
 end
 
 get '/memos/:memo_id/edit' do
-  database = MemoDatabase.new
-  @memo = database.search_memo_by_id(params['memo_id'])
+  @memo = settings.database.search_memo_by_id(params['memo_id'])
   @id = params['memo_id']
   @title = 'Edit memo'
   pass unless @memo
@@ -54,80 +56,28 @@ get '/memos/:memo_id/edit' do
 end
 
 post '/memos/' do
-  database = MemoDatabase.new
-  database.add(title: params[:title], content: params[:content])
+  settings.database.add(title: params[:title], content: params[:content])
 
   redirect '/memos'
 end
 
 delete '/memos/:memo_id' do
-  database = MemoDatabase.new
-  memo = database.search_memo_by_id(params['memo_id'])
+  memo = settings.database.search_memo_by_id(params['memo_id'])
   id = params['memo_id']
   pass if memo.nil?
 
-  database.delete(id)
+  settings.database.delete(id)
   redirect '/memos'
 end
 
 patch '/memos/:memo_id' do
-  database = MemoDatabase.new
   id = params['memo_id']
 
-  database.update(id: id, title: params['title'], content: params['content'])
+  settings.database.update(id: id, title: params['title'], content: params['content'])
   redirect '/memos'
 end
 
 not_found do
   status 404
   '404 - Not Found'
-end
-
-# メモの読み書き、表示、削除用のクラス。
-class MemoDatabase
-  attr_accessor :memos
-
-  OUTPUT_PATH = './output/'
-  CSV_PATH = './output/Sample.csv'
-
-  def initialize
-    unless File.exist?(CSV_PATH)
-      Dir.mkdir('output', 0o755) unless File.exist?(OUTPUT_PATH)
-      File.write(CSV_PATH, <<~CSV)
-        id,title,contetnt
-      CSV
-    end
-
-    @memos = CSV.read(CSV_PATH, headers: true)
-  end
-
-  def load_all_memos
-    @memos
-  end
-
-  def search_memo_by_id(id)
-    @memos.detect { |row| row['id'] == id }
-  end
-
-  def write
-    File.write(CSV_PATH, @memos)
-  end
-
-  def add(title:, content:)
-    new_id = SecureRandom.uuid
-    @memos << CSV::Row.new(%w[id title content], [new_id, title, content])
-    write
-  end
-
-  def update(id:, title:, content:)
-    row_index = @memos['id'].find_index(id)
-    @memos[row_index]['title'] = title
-    @memos[row_index]['content'] = content
-    write
-  end
-
-  def delete(id)
-    @memos = CSV::Table.new(@memos.reject { |row| row['id'] == id })
-    write
-  end
 end
